HTML extract of the report.
You can find an extract from Jupyter Notebook with the code and example output (the information is anonymized).

This code parses data to identify suppliers with lacking recognition rates, good recognition, as well as those for which we still lack information. It also produces infographics for better visulization of achievements and areas for improvement.